Fix TrajOptIfoptOSQPSolverProfile constructors and move opt_params to parent class
